1、优化滚动条的滚动块长度计算，设置最小长度； 2、解决listctrl、grdictrl、treectrl控件的滚动条位置计算的问题。

+++ DuiVision/source/DuiTreeCtrl.cpp
@@ -1708,7 +1708,7 @@ void CDuiTreeCtrl::DrawControl(CDC &dc, CRect rcUpdate)
 	int nCurPos = pScrollV->GetScrollCurrentPos();	// 当前top位置
 	int nMaxRange = pScrollV->GetScrollMaxRange();
 
-	m_nVirtualTop = (nMaxRange > 0) ? nCurPos*(nHeightAll-m_rc.Height())/nMaxRange : 0;	// 当前滚动条位置对应的虚拟的top位置
+	m_nVirtualTop = (nMaxRange > 0) ? (int)((double)nCurPos*(nHeightAll-m_rc.Height())/nMaxRange) : 0;	// 当前滚动条位置对应的虚拟的top位置
 	if(m_nVirtualTop < 0)
 	{
 		m_nVirtualTop = 0;
